Changed how default is displayed in SPA.
Fixed null reference errors in SPA around step and step category.

Added autofill to OE screen customer input.

Added some model data and data access props to OE to get ready to start working more with it soon.

+++ Armis.Data/DatabaseContext/Entities/Part.cs
@@ -28,7 +28,7 @@ namespace Armis.Data.DatabaseContext.Entities
         public decimal? BasePrice { get; set; }
         public decimal? MinLotCharge { get; set; }
         public int? PartsPerLoad { get; set; }
-        public int? MaskPcsPerHour { get; set; }
+        public decimal? MaskPcsPerHour { get; set; }
         public bool? NotifyWhenMasking { get; set; }
         public int? Alloy { get; set; }
         public int? SeriesId { get; set; }
